Rename and fix test
